+++ core/src/main/java/io/questdb/griffin/SqlParser.java
@@ -2524,7 +2524,7 @@ public class SqlParser {
     }
 
     private int toColumnType(GenericLexer lexer, CharSequence tok) throws SqlException {
-        final short type = ColumnType.tagOf(tok);
+        final byte type = ColumnType.tagOf(tok);
         if (type == -1) {
             throw SqlException.$(lexer.lastTokenPosition(), "unsupported column type: ").put(tok);
         }
